+++ modules/ui/misc/index.tsx
@@ -1,4 +1,4 @@
 export * from "./Catcher.js";
-export * from "./ElementMap.js";
+export * from "./ElementMapper.js";
 export * from "./Loading.js";
 export * from "./Meta.js";
