+++ modules/ui/tree/TreeCards.tsx
@@ -1,6 +1,6 @@
 import type { FunctionComponent, ReactNode } from "react";
 import { type Element, type ElementProps, type Elements, getElements } from "../../util/element.js";
-import { mapElements } from "../misc/ElementMap.js";
+import { mapElements } from "../misc/ElementMapper.js";
 import TREE_CARDS_CSS from "./TreeCards.module.css";
 
 export interface TreeCardsProps {
